requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
pytest==7.4.3
Flask-Login==0.6.3
psycopg2-binary==2.9.9
flask_wtf==1.1.1
is_safe_url==1.0
matplotlib==3.5.3
gunicorn
werkzeug==2.2.3